requirements.txt

Django==2.0.3
django-ckeditor==5.6.1
django-cleanup==3.1.0
django-hitcount==1.3.0
Pillow== 5.3
